backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
httpx==0.26.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==1.0.0
apscheduler==3.10.4
jinja2==3.1.3
pydantic-settings==2.1.0
toml==0.10.2
pyyaml==6.0.1

